Adding C Compiler/Linking Practice Code:
Adding the code I used for playing with the
C preproccessor, compiler, and linking phases
for building an executable. It starts with basic
single-file compilation and ends with building
shared libraries.

It's an interesting exercise for checking out
how the preprocessor fills in macros and to see
what libraries are linked with an executable or
library.

Signed-off-by: mrandronov <mrandronov88@gmail.com>