Adding Christmass Tree Printer:
This is a test program to get in the mindset of
C syntax. The goal of the program is to basically
print a triangle (christmas tree ;) ) in all four
cardinal directions. It's a fun program to
implement that gets you thinking about managing
loops and variables in the like in a reasonably
non-trivial way.

Signed-off-by: mrandronov <mrandronov88@gmail.com>