Adding Morse Code Translator exercise:
This is a larger exercise, where the goal is to
write a program that will translate English into
Morse code. The implementation details are
described in the code, but I figured the most
maintainable way to write this was by splitting
parts of the required information for translation
into separate header files.

I think this one can be further built upon, so
perhaps I will revisit it. This was a fun
exercise to do and I think it helps with thinking
of a solution in a 'C' way.

Signed-off-by: mrandronov <mrandronov88@gmail.com>