Adding Table of Contents Exercise:
This was an interesting problem I encountered
during my job hunt. It is to generate a table of
contents from text that contains markdown syntaxs
for headings. I found it hard to solve, but good
practice for separating the implementations and
definitions, as well as creating libraries, and
building out the Makefile in a reasonable way.
I don't think it is big enough to be it's own
project, so I think it fits well as an C exercise.

Signed-off-by: mrandronov <mrandronov88@gmail.com>